주문내역 페이지에서 주문테이블에있는 address로 들고 오도록 변경

+++ LaClave/src/main/java/com/itwillbs/LaClave/wishlist/Wishlist.java
@@ -26,7 +26,7 @@ public class Wishlist {
     private Integer wishlistIdx;
 
     @Column(name = "MEMBER_IDX", nullable = false)
-    private Integer memberIdx;
+    private Long memberIdx;
 
     @Column(name = "PRODUCT_IDX", nullable = false)
     private Integer productIdx;
