+++ LaClave/src/main/java/com/itwillbs/LaClave/wishlist/WishlistRepository.java
@@ -10,7 +10,7 @@ public interface WishlistRepository extends JpaRepository<Wishlist, Integer>{
 	List<Wishlist> findByMemberIdxOrderByWishlistDateDesc(Integer memberIdx);
 	
 	//위시 리스트 중복 체크
-	boolean existsByMemberIdxAndProductIdx(Integer memberIdx, Integer productIdx);
+	boolean existsByMemberIdxAndProductIdx(Long memberIdx, Integer productIdx);
 	
 	/* 상품 기준 삭제 (옵션) */  
 	// 하트 취소 버튼용
